# Initial support for clique-based kernels