Support for clique-based kernels to take advantage of multi-linked XGMI-based topologies.

- This is currently an experimental feature that needs to be opted into via the environment variable RCCL_ENABLE_CLIQUE=1 (on all ranks).
- Clique-based kernels currently launch via a different mechanism and is not compatible with ncclGroupStart / ncclGroupEnd at this time.
- Only AllReduce clique-based kernels are supported at this time.
- A sample HelloRccl program has been added to tools to compare performance between the current algorithms / clique-based kernels.
- Clique-based kernels are expected to be faster for smaller input sizes, but slower for larger ones

Sample output:
```
With clique-based kernels:
AllReduce Performance (sum of floats):
     Bytes CpuTime(ms) GpuTime(ms)
      4096      0.015      0.014
      8192      0.015      0.015
     16384      0.015      0.014
     32768      0.017      0.016
     65536      0.017      0.017
    131072      0.018      0.017
    262144      0.020      0.019
    524288      0.027      0.026
   1048576      0.053      0.053
   2097152      0.119      0.119
   4194304      0.250      0.250
   8388608      0.477      0.476
  16777216      0.921      0.921
  33554432      1.768      1.768
  67108864      3.520      3.520
 134217728      7.157      7.152
 268435456     14.009     14.008
 536870912     28.134     28.133
1073741824     55.917     55.912
Without clique-based kernels:
AllReduce Performance (sum of floats):
     Bytes CpuTime(ms) GpuTime(ms)
      4096      0.059      0.058
      8192      0.060      0.059
     16384      0.061      0.061
     32768      0.065      0.065
     65536      0.072      0.072
    131072      0.096      0.096
    262144      0.073      0.073
    524288      0.084      0.083
   1048576      0.109      0.109
   2097152      0.156      0.156
   4194304      0.260      0.260
   8388608      0.443      0.443
  16777216      0.793      0.793
  33554432      1.599      1.599
  67108864      2.998      2.998
 134217728      5.810      5.810
 268435456     12.263     12.262
 536870912     22.436     22.436
1073741824     44.619     44.618

```