@wenkaidu 2020-10-06
1. Many files are missing copyrights header
2. What is the algorithm of AUTOUNROLL calculation? Can it fall back to (UNROLL*(4/(MINDSTS+MINSRCS))) in case of non-clique kernels to avoid performance re-evaluation?