@gilbertlee-amd 2020-10-07
> 1. Many files are missing copyrights header
> 2. What is the algorithm of AUTOUNROLL calculation? Can it fall back to (UNROLL*(4/(MINDSTS+MINSRCS))) in case of non-clique kernels to avoid performance re-evaluation?

1) Thanks - Added
2) I've fixed it so that it should be equivalent with the original intent.  With the clique-based kernels, MINDSTS+MINSRCS can be greater than 4, which was resulting in AUTOUNROLL set to 0.